+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.2 [2022-02-25]
+- BREAKING: changes parameters of SlideCallback
+- optimizes fade animation
+
 ## 0.1.1 [2022-02-21]
 - fixes README.md